fix: component splitting and refactoring
