+++ web/app/[workspaceSlug]/(projects)/projects/(detail)/archives/page.tsx
@@ -1,0 +1,4 @@
+import ProjectPageRoot from "@/plane-web/components/projects/page";
+
+const ProjectsPage = () => <ProjectPageRoot />;
+export default ProjectsPage;
