+++ web/ce/types/projects/index.ts
@@ -1,0 +1,1 @@
+export * from "./projects";
