+++ web/core/components/gantt-chart/blocks/blocks-list.tsx
@@ -14,10 +14,10 @@ export type GanttChartBlocksProps = {
   getBlockById: (id: string, currentViewData?: ChartDataType | undefined) => IGanttBlock;
   blockToRender: (data: any) => React.ReactNode;
   blockUpdateHandler: (block: any, payload: IBlockUpdateData) => void;
-  enableBlockLeftResize: boolean;
-  enableBlockRightResize: boolean;
-  enableBlockMove: boolean;
-  enableAddBlock: boolean;
+  enableBlockLeftResize: boolean | ((blockId: string) => boolean);
+  enableBlockRightResize: boolean | ((blockId: string) => boolean);
+  enableBlockMove: boolean | ((blockId: string) => boolean);
+  enableAddBlock: boolean | ((blockId: string) => boolean);
   ganttContainerRef: React.RefObject<HTMLDivElement>;
   showAllBlocks: boolean;
   selectionHelpers: TSelectionHelper;
@@ -55,10 +55,14 @@ export const GanttChartBlocksList: FC<GanttChartBlocksProps> = (props) => {
           showAllBlocks={showAllBlocks}
           blockToRender={blockToRender}
           blockUpdateHandler={blockUpdateHandler}
-          enableBlockLeftResize={enableBlockLeftResize}
-          enableBlockRightResize={enableBlockRightResize}
-          enableBlockMove={enableBlockMove}
-          enableAddBlock={enableAddBlock}
+          enableBlockLeftResize={
+            typeof enableBlockLeftResize === "function" ? enableBlockLeftResize(blockId) : enableBlockLeftResize
+          }
+          enableBlockRightResize={
+            typeof enableBlockRightResize === "function" ? enableBlockRightResize(blockId) : enableBlockRightResize
+          }
+          enableBlockMove={typeof enableBlockMove === "function" ? enableBlockMove(blockId) : enableBlockMove}
+          enableAddBlock={typeof enableAddBlock === "function" ? enableAddBlock(blockId) : enableAddBlock}
           ganttContainerRef={ganttContainerRef}
           selectionHelpers={selectionHelpers}
         />
