+++ web/core/components/project/create-project-modal.tsx
@@ -2,6 +2,7 @@ import { useEffect, Fragment, FC, useState } from "react";
 import { Dialog, Transition } from "@headlessui/react";
 // components
 import { CreateProjectForm } from "@/plane-web/components/projects/create/root";
+import { TProject } from "@/plane-web/types/projects";
 import { ProjectFeatureUpdate } from "./project-feature-update";
 
 type Props = {
@@ -9,6 +10,7 @@ type Props = {
   onClose: () => void;
   setToFavorite?: boolean;
   workspaceSlug: string;
+  data?: Partial<TProject>;
 };
 
 enum EProjectCreationSteps {
@@ -17,7 +19,7 @@ enum EProjectCreationSteps {
 }
 
 export const CreateProjectModal: FC<Props> = (props) => {
-  const { isOpen, onClose, setToFavorite = false, workspaceSlug } = props;
+  const { isOpen, onClose, setToFavorite = false, workspaceSlug, data } = props;
   // states
   const [currentStep, setCurrentStep] = useState<EProjectCreationSteps>(EProjectCreationSteps.CREATE_PROJECT);
   const [createdProjectId, setCreatedProjectId] = useState<string | null>(null);
@@ -68,6 +70,7 @@ export const CreateProjectModal: FC<Props> = (props) => {
                     workspaceSlug={workspaceSlug}
                     onClose={onClose}
                     handleNextStep={handleNextStep}
+                    data={data}
                   />
                 )}
                 {currentStep === EProjectCreationSteps.FEATURE_SELECTION && (
