+++ web/core/components/workspace/sidebar/workspace-menu.tsx
@@ -1,13 +1,13 @@
 "use client";
 
-import React, { useEffect } from "react";
+import React, { useEffect, useRef, useState } from "react";
 import { observer } from "mobx-react";
 import Link from "next/link";
 import { useParams, usePathname } from "next/navigation";
-import { ChevronRight } from "lucide-react";
+import { ArchiveIcon, ChevronRight, MoreHorizontal, Settings } from "lucide-react";
 import { Disclosure, Transition } from "@headlessui/react";
 // ui
-import { Tooltip } from "@plane/ui";
+import { CustomMenu, Tooltip } from "@plane/ui";
 // components
 import { SidebarNavItem } from "@/components/sidebar";
 // constants
@@ -19,11 +19,16 @@ import { cn } from "@/helpers/common.helper";
 // hooks
 import { useAppTheme, useEventTracker, useUser } from "@/hooks/store";
 import useLocalStorage from "@/hooks/use-local-storage";
+import useOutsideClickDetector from "@/hooks/use-outside-click-detector";
 import { usePlatformOS } from "@/hooks/use-platform-os";
 // plane web components
 import { UpgradeBadge } from "@/plane-web/components/workspace";
 
 export const SidebarWorkspaceMenu = observer(() => {
+  // state
+  const [isMenuActive, setIsMenuActive] = useState(false);
+  // refs
+  const actionSectionRef = useRef<HTMLDivElement | null>(null);
   // store hooks
   const { toggleSidebar, sidebarCollapsed } = useAppTheme();
   const { captureEvent } = useEventTracker();
@@ -54,30 +59,75 @@ export const SidebarWorkspaceMenu = observer(() => {
   useEffect(() => {
     if (sidebarCollapsed) toggleWorkspaceMenu(true);
   }, [sidebarCollapsed, toggleWorkspaceMenu]);
-
-  const indicatorElement = (
-    <div className="flex-shrink-0">
-      <UpgradeBadge />
-    </div>
-  );
+  useOutsideClickDetector(actionSectionRef, () => setIsMenuActive(false));
 
   return (
     <Disclosure as="div" defaultOpen>
       {!sidebarCollapsed && (
-        <Disclosure.Button
-          as="button"
-          className="sticky top-0 bg-custom-sidebar-background-100 z-10 group/workspace-button w-full px-2 py-1.5 flex items-center justify-between gap-1 text-custom-sidebar-text-400 hover:bg-custom-sidebar-background-90 rounded text-xs font-semibold"
-          onClick={() => toggleWorkspaceMenu(!isWorkspaceMenuOpen)}
-        >
-          <span>WORKSPACE</span>
-          <span className="flex-shrink-0 opacity-0 pointer-events-none group-hover/workspace-button:opacity-100 group-hover/workspace-button:pointer-events-auto rounded p-0.5 hover:bg-custom-sidebar-background-80">
-            <ChevronRight
-              className={cn("size-4 flex-shrink-0 text-custom-sidebar-text-400 transition-transform", {
-                "rotate-90": isWorkspaceMenuOpen,
-              })}
-            />
-          </span>
-        </Disclosure.Button>
+        <div className="flex px-2 bg-custom-sidebar-background-100 group/workspace-button hover:bg-custom-sidebar-background-90 rounded">
+          {" "}
+          <Disclosure.Button
+            as="button"
+            className="flex-1 sticky top-0  z-10  w-full  py-1.5 flex items-center justify-between gap-1 text-custom-sidebar-text-400  text-xs font-semibold"
+            onClick={() => toggleWorkspaceMenu(!isWorkspaceMenuOpen)}
+          >
+            <span>WORKSPACE</span>
+          </Disclosure.Button>
+          <CustomMenu
+            customButton={
+              <span
+                ref={actionSectionRef}
+                className="grid place-items-center p-0.5 text-custom-sidebar-text-400 hover:bg-custom-sidebar-background-80 rounded my-auto"
+                onClick={() => {
+                  console.log("ndkn");
+                  setIsMenuActive(!isMenuActive);
+                }}
+              >
+                <MoreHorizontal className="size-4" />
+              </span>
+            }
+            className={cn(
+              "h-full flex items-center opacity-0 z-20 pointer-events-none flex-shrink-0 group-hover/workspace-button:opacity-100 group-hover/workspace-button:pointer-events-auto my-auto",
+              {
+                "opacity-100 pointer-events-auto": isMenuActive,
+              }
+            )}
+            customButtonClassName="grid place-items-center"
+            placement="bottom-start"
+          >
+            <CustomMenu.MenuItem>
+              <Link href={`/${workspaceSlug}/projects/archives`}>
+                <div className="flex items-center justify-start gap-2">
+                  <ArchiveIcon className="h-3.5 w-3.5 stroke-[1.5]" />
+                  <span>Archives</span>
+                </div>
+              </Link>
+            </CustomMenu.MenuItem>
+
+            <CustomMenu.MenuItem>
+              <Link href={`/${workspaceSlug}/settings`}>
+                <div className="flex items-center justify-start gap-2">
+                  <Settings className="h-3.5 w-3.5 stroke-[1.5]" />
+                  <span>Settings</span>
+                </div>
+              </Link>
+            </CustomMenu.MenuItem>
+          </CustomMenu>
+          <Disclosure.Button
+            as="button"
+            className="sticky top-0 z-10 group/workspace-button px-0.5 py-1.5 flex items-center justify-between gap-1 text-custom-sidebar-text-400 hover:bg-custom-sidebar-background-90 rounded text-xs font-semibold"
+            onClick={() => toggleWorkspaceMenu(!isWorkspaceMenuOpen)}
+          >
+            {" "}
+            <span className="flex-shrink-0 opacity-0 pointer-events-none group-hover/workspace-button:opacity-100 group-hover/workspace-button:pointer-events-auto rounded hover:bg-custom-sidebar-background-80">
+              <ChevronRight
+                className={cn("size-4 flex-shrink-0 text-custom-sidebar-text-400 transition-transform", {
+                  "rotate-90": isWorkspaceMenuOpen,
+                })}
+              />
+            </span>
+          </Disclosure.Button>
+        </div>
       )}
       <Transition
         show={isWorkspaceMenuOpen}
@@ -121,7 +171,11 @@ export const SidebarWorkspaceMenu = observer(() => {
                           />
                           {!sidebarCollapsed && <p className="text-sm leading-5 font-medium">{link.label}</p>}
                         </div>
-                        {!sidebarCollapsed && link.key === "active-cycles" && indicatorElement}
+                        {!sidebarCollapsed && link.key === "active-cycles" && (
+                          <div className="flex-shrink-0">
+                            <UpgradeBadge flag="WORKSPACE_ACTIVE_CYCLES" />
+                          </div>
+                        )}
                       </SidebarNavItem>
                     </Link>
                   </Tooltip>
